pyproject.toml
[tool.poetry]
name = "autobot-ml"
version = "0.0.1"
description = "An automated code refactoring tool powered by GPT-3."
authors = ["Charlie Marsh <charlie.r.marsh@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/charliermarsh/autobot"
keywords = ["automation", "refactor", "GPT-3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
packages = [{ include = "autobot" }]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
colorama = "^0.4.5"
openai = "^0.23.0"
python-dotenv = "^0.21.0"
rich = "^12.5.1"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"
mypy = "^0.971"

[tool.poetry.scripts]
autobot = "autobot.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
extend-exclude = "schematics"
preview = true

[tool.isort]
profile = "black"
extend_skip = "schematics"
